backend/requirements.txt
# Market Intelligence Dashboard - Requirements
# Python 3.10+

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Database
sqlalchemy[asyncio]==2.0.25
aiosqlite==0.19.0
alembic==1.13.1

# LLM - Z.AI GLM-4.7 via OpenAI-compatible API
openai>=1.0

# Scheduling
apscheduler==3.10.4

# RSS/XML Parsing
feedparser==6.0.10
beautifulsoup4==4.12.3
lxml==5.1.0

# Data Processing
python-dateutil==2.8.2
pydantic==2.5.3
pydantic-settings==2.1.0

# PDF Processing
pymupdf==1.24.0

# Utilities
python-dotenv==1.0.0
loguru==0.7.2

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
